Move OpenCVTypeAdapters to core-processing
Addresses https://github.com/qupath/qupath/issues/961
