+++ qupath-core/src/main/java/qupath/lib/io/GsonTools.java
@@ -76,7 +76,6 @@ public class GsonTools {
 			.serializeSpecialFloatingPointValues()
 			.setLenient()
 			.registerTypeAdapterFactory(new QuPathTypeAdapterFactory())
-			.registerTypeAdapterFactory(OpenCVTypeAdapters.getOpenCVTypeAdaptorFactory())
 			.registerTypeAdapter(AffineTransform.class, AffineTransformTypeAdapter.INSTANCE);
 			//.create();
